feat: improves page structures
